ensure javascript files are joined with a semicolon
This is a potential problem introduced by reversing
the order in which javascript files are concattenated
and minified.

Before, the logic was like this:

    data = read_file(f) for f in paths
    concatted = concat(data)
    minified = minify(concatted)

Now, since we are making decisions *per path* on whether
to minify or not, we're doing something like this:

    minified_data = minify(path) for f in paths
    minified = concat(minified_data)

The problem is that many javascript minifiers, which rely
heavily on ASI, will strip the final semicolon off the end
of the output.  This can present a problem, for example:

    one.js:
      alert('one!');

    two.js:
      (function() { alert('two!'); })();

will result in:

    one-two.min.js:
      alert('one!')(function(){alert('two!')})()

and give you a nasty "undefined is not callable" error
which is really tricky to track down.

The solution is to always join javascript files with semicolons
rather than empty strings.  This way we'd end up with:

    one-two.min.js:
      alert('one!');(function(){alert('two!')})()

which does what you expect.

for more info, see here: https://github.com/mishoo/UglifyJS/issues/126